+++ Lib/Logic_layer/LL.h
@@ -9,8 +9,8 @@
 /* INCLUDES ******************************/
 #include <stdio.h>
 #include "main.h"
-#include "IO_layer_Lib/IO_func.h"
-#include "Parser/parser.h"
+#include "IO_layer/IO_func.h"
+#include "Front_layer/parser.h"
 /* DEFINES ******************************/
 #define AMOUNTOFCOLORS	 17		/*!< amount of colors that are recognized */
 #define SYNTAX_ERROR_COLOR 0x01 // internal error code
@@ -22,5 +22,5 @@ typedef struct command COMMAND;
 /* EXTERN VARIABLES ******************************/
 
 /* PROTOTYPES ******************************/
-uint8_t LL_textToColor(char *text);
-uint8_t LL_executeCommand(COMMAND *c, uint8_t last_place);
+uint16_t LL_textToColor(char *text);
+uint16_t LL_executeCommand(COMMAND *c, uint8_t last_place);
